Fix item functions
