+++ route/itemRoute.js
@@ -1,38 +1,23 @@
 const express = require("express");
 const db = require("../database/db_item");
+const fn = require('../function/functions');
 const app = express.Router()
 
-function isItemIdExist(db, itemId) {
-    const result = db.find(({ id }) => id == itemId)
-    return result
-}
+
 
 app.get('/item/', (req, res) => {
     res.send(db)
 })
 
 app.get('/item/:id', (req, res) => {
-    const id = isItemIdExist(db, req.params.id)
+    const id = fn.isItemIdExist(db, req.params.id)
     id ? res.status(200).send(id) : res.status(404).send('item does not exist')
 })
 
-function isIdExist(id) {
-    if (db.some(db => db.id == id)) return true;
-    else return false;
-}
-
-function isItemDataComplete(itemdb){
-    if('id' in itemdb
-        && 'userId' in itemdb
-        && 'name' in itemdb
-    ) return true;
-    else return false;
-}
-
 app.post('/item/', (req, res) => {
-    if (isIdExist(req.body.id)) {
+    if (fn.isIdExist(db, req.body.id)) {
         res.send('Id is already use')
-    } else if (isItemDataComplete(req.body) == false){
+    } else if (fn.isItemDataComplete(req.body) == false){
         res.send('Complete the data')
     } else {
         db.push(req.body)
@@ -46,9 +31,9 @@ app.put('/item/:index', (req, res) => {
         res.status(400).send('Input number please')
     } else if ((db.length - 1) < index) {
         res.status(400).send('Cannot found number of item')
-    } else if (isIdExist(req.body.id)) {
+    } else if (fn.isIdExist(db, req.body.id)) {
         res.send('Id is already use')
-    } else if (isItemDataComplete(req.body) == false){
+    } else if (fn.isItemDataComplete(req.body) == false){
         res.send('Complete the data')
     } else {
         db[req.params.index] = req.body
@@ -61,5 +46,7 @@ app.delete('/item/:index', (req, res) => {
     const deletedItem = db.splice(index, 1)
     res.send(deletedItem)
 })
+
+
 
 module.exports = app
